+++ app/src/main/java/eparon/nxtremotecontroller/NXT/Motor.java
@@ -1,0 +1,14 @@
+package eparon.nxtremotecontroller.NXT;
+
+public class Motor {
+
+    public static final byte NXT_A = (byte)0x00;
+    public static final byte NXT_B = (byte)0x01;
+    public static final byte NXT_C = (byte)0x02;
+
+    public static final byte EV3_A = (byte)1;
+    public static final byte EV3_B = (byte)2;
+    public static final byte EV3_C = (byte)4;
+    public static final byte EV3_D = (byte)8;
+
+}